requirements.txt
celery[redis]>=5.4,<6.0
flower>=2.0,<3.0
alembic>=1.13,<2.0
playwright>=1.45,<2.0
SQLAlchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
pydantic>=2.8,<3.0
pydantic-settings>=2.4,<3.0
requests>=2.32,<3.0
python-dateutil>=2.9,<3.0
beautifulsoup4>=4.12,<5.0
tenacity>=8.5,<10.0
python-dotenv>=1.0,<2.0
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
